# gc-spends-backend/docs/swagger/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "GC Spends API",
    "version": "1.0.0",
    "description": "GC Spends API - Система управления расходами и платежами\n\nЭто API для управления системой расходов и платежей, включающая в себя:\n\n* **Аутентификация и авторизация** - управление пользователями и ролями\n* **Управление запросами** - создание и обработка платежных запросов\n* **Справочники** - управление статьями расходов и другими справочными данными\n* **Регистрация** - ведение реестра документов\n* **Распределение** - управление распределением средств\n* **Экспорт контрактов** - работа с контрактными данными\n* **Файлы** - загрузка и управление файлами\n* **Приоритеты** - управление приоритетами платежных запросов\n* **Мониторинг** - мониторинг и наблюдение за системой\n* **Идемпотентность** - управление идемпотентными ключами\n\n### Аутентификация\nAPI использует JWT токены для аутентификации. Получите токен через `/api/v1/auth/login` и используйте его в заголовке `Authorization: Bearer <token>`.\n\n### Базовый URL\nВсе API endpoints доступны по адресу: `http://localhost:8000/api/v1/`",
    "contact": {
      "name": "GC Spends Team",
      "email": "support@gcspends.com"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "https://gcback.openlayers.kz",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/api/v1/priority/rules": {
      "post": {
        "tags": [
          "priority"
        ],
        "summary": "Создать правило приоритета",
        "description": "Создает новое правило для расчета приоритета платежных запросов",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PriorityRuleCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Правило создано",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PriorityRuleOut"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "priority"
        ],
        "summary": "Получить список правил приоритета",
        "description": "Возвращает список правил для расчета приоритета",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "description": "Количество записей для пропуска",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Максимальное количество записей",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000,
              "default": 100
            }
          },
          {
            "name": "active_only",
            "in": "query",
            "description": "Показать только активные правила",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список правил",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "rules": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PriorityRuleOut"
                      }
                    },
                    "total": {
                      "type": "integer"
                    },
                    "skip": {
                      "type": "integer"
                    },
                    "limit": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/sub-registrar/assignments": {
      "get": {
        "tags": [
          "sub-registrar"
        ],
        "summary": "Получить назначения суб-регистратора",
        "description": "Возвращает все назначения для текущего суб-регистратора",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "description": "Количество записей для пропуска",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Максимальное количество записей",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000,
              "default": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Список назначений",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubRegistrarAssignmentListOut"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          }
        }
      }
    },
    "/api/v1/sub-registrar/reports/{request_id}": {
      "get": {
        "tags": [
          "sub-registrar"
        ],
        "summary": "Получить отчет суб-регистратора",
        "description": "Возвращает отчет суб-регистратора для конкретного запроса",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "description": "ID запроса",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Отчет суб-регистратора",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubRegistrarReportOut"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenError"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "string"
          }
        }
      },
...
      "PriorityRuleCreate": {
        "type": "object",
        "required": [
          "name",
          "priority",
          "conditions"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "priority": {
            "type": "string",
            "enum": [
              "low",
              "normal",
              "high",
              "urgent",
              "critical"
            ]
          },
          "conditions": {
            "type": "object",
            "description": "Условия для применения правила"
          },
          "is_active": {
            "type": "boolean",
            "default": true
          }
        }
      },
      "PriorityRuleOut": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "priority": {
            "type": "string",
            "enum": [
              "low",
              "normal",
              "high",
              "urgent",
              "critical"
            ]
          },
          "conditions": {
            "type": "object"
          },
          "is_active": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "SubRegistrarAssignmentListOut": {
        "type": "object",
        "properties": {
          "assignments": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "total": {
            "type": "integer"
          }
        }
      },
      "SubRegistrarReportOut": {
        "type": "object",
        "properties": {
          "request_id": {
            "type": "string",
            "format": "uuid"
          },
          "report_data": {
            "type": "object"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "responses": {
      "NotFoundError": {
        "description": "Ресурс не найден",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "ForbiddenError": {
        "description": "Доступ запрещен",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  }
}